python/pyproject.toml
[project]
name = "neo4j-mcp-ecommerce"
version = "0.1.0"
description = "MCP server for an ecommerce store backed by Neo4j"
authors = [{ name = "Neo4j MCP Workshop" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "mcp>=1.1.0",
  "neo4j>=5.22.0",
  "python-dotenv>=1.0.1",
  "anyio>=4.4.0",
  "pydantic>=2.7.0",
  "uvicorn>=0.30.0",
  "fastapi>=0.110.0",
]

[project.scripts]
neo4j-mcp-ecommerce = "neo4j_mcp.main:run"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/neo4j_mcp"]

[tool.uv]
# uv project configuration
